requirements.txt
numpy==1.26.4
kivy
kivymd
git+https://github.com/FacePerceiver/facer.git@main
opencv-contrib-python
jax==0.5.0
jax-metal
colour-science
scikit-image
rawpy
lensfunpy
pyexiftool
setuptools
multipledispatch
wheel
torch
torchvision
wand
watchfiles
screeninfo
pyopencl
numba
thop
einops
timm
matplotlib
git+https://github.com/google-ai-edge/mediapipe.git@main
typer
typer_config
fastapi
omegaconf
etils
huggingface_hub
pillow_heif